refactor: gather call signature text and markup in one module
The text of a signature was built in a module that also runs ruff,
formats names and renders reprs; the markup was built by module-level
functions sitting in front of a render mixin that never called them on
itself. Both now live together, and the mixin builds its lines and hands
them over.
